% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.R
\name{geoLet}
\alias{geoLet}
\title{class for loading and presenting DICOM data}
\usage{
geoLet(use.ROICache = FALSE)
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname,
              methods are exposed with the technique of 'closure'.
              In order to see manuals for the single mathods, consider the vignette or use the
              available for the following wrapping functions:
              \itemize{
              \item \code{openDICOMFolder( );} : to load a DICOM series into an geoLet object
              \item \code{getImageVoxelCube( );} : to get the ImageVoxelCube stored into a geoLet object
              \item \code{getPixelSpacing( );} : to get the pixelSpacing (x,y,z) of the main ImageVoxelCube stored into a geoLet object
              \item \code{getROIList( );} : to get the list of the ROI defined in a geoLet object
              \item \code{getTag( );} : to get a single DICOM-tag of a DICOM file loaded into a geoLet object
              \item \code{getROIVoxels( );} : to get the IMAGE Voxels geometrically located into a ROI, for a given geoLet object
              \item \code{extractDoseVoxels( );} : to get the DOSE Voxels geometrically located into a ROI, for a given geoLet object
              \item \code{calculateDVH( );} : to get the DVH calculated from a geoLet object
              }
              The original methods for the class geoLet can also be invocked using the same name without the previx 'GTL.', i.e.:
              misc3d rgl Rvcg oce rmarkdown moments
}
